test: move test/test-lib to test-lib. Lint dislikes test imports
