+++ a3p-integration/proposals/n:upgrade-next/test-lib/price-feed.js
